+++ alias.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+alias etlas='../etlas'