+++ src/main/eta/helvm/helcam/BrainFuck/Token.hs
@@ -32,4 +32,3 @@ instance Read Token where
   readsPrec _ "[" = [( JmpPast, "")]
   readsPrec _ "]" = [( JmpBack, "")]
   readsPrec _ _   = []
-