+++ src/main/eta/helvm/helcam/Common/Util.hs
@@ -1,0 +1,6 @@
+module HelVM.HelCam.Common.Util where
+
+charToString :: Char -> String
+charToString = (:[])
+
+type D a = a -> a