+++ src/main/eta/helvm/helcam/WhiteSpace/Tokens.hs
@@ -1,7 +1,6 @@
-module HelVM.HelCam.BrainFuck.Tokens where
+module HelVM.HelCam.WhiteSpace.Tokens where
 
-import HelVM.HelCam.BrainFuck.Token
-import HelVM.HelCam.Common.Util
+import HelVM.HelCam.WhiteSpace.Tokens
 
 import Data.Maybe
 import Text.Read
@@ -13,6 +12,7 @@ newtype Tokens = Tokens TokenList
 instance Show Tokens where
   show (Tokens tokens) = tokens >>= show
 
+-- Lexer
 instance Read Tokens where
   readsPrec _ source = [( Tokens $ source >>= maybeToList . readMaybe . charToString, "")]
 
@@ -24,6 +24,6 @@ tokenList (Tokens tokens) = tokens
 readTokens :: String -> Tokens
 readTokens source = read source :: Tokens
 
--- Lexer
 tokenize :: String -> TokenList
 tokenize =  tokenList . readTokens
+