+++ src/test/eta/Test.hs
@@ -1,6 +1,7 @@
 module Main(main) where
 
 import HelVM.HelCam.BrainFuck.TokensTest
+import HelVM.HelCam.WhiteSpace.ParserTest
 
 import Test.HUnit
 
@@ -8,9 +9,11 @@ testExample :: Test
 testExample = TestCase (assertEqual "test" "test" "test")
 
 testList :: Test
-testList = TestList [ TestLabel "testExample" testExample
-                    , TestLabel "testsOfTokens" testsOfTokens
-                    ]
+testList = TestList 
+  [ TestLabel "testExample" testExample
+  , TestLabel "testsOfTokens" testsOfTokens
+  , TestLabel "testOfParser" testsOfParser
+  ]
 
 main :: IO ()
 main = do