+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for HelCam
 
+## 0.3.0.0  -- 2020-05-30
+
+* Add WhiteSpace Parser
+
 ## 0.2.0.0  -- 2020-05-16
 
 * Add BrainFuck Lexer