+++ docs/NEWS.md
@@ -2,6 +2,7 @@
 
 **New features.**
 
-* Add BrainFuck Lexer
+* Add Parser for WhiteSpace
+* Add Lexer for BrainFuck 
 
 For more see [CHANGELOG](CHANGELOG.md).